Move GroupData and associated types to a separate file
