+++ perf-event/src/lib.rs
@@ -78,6 +78,7 @@
 pub mod events;
 
 mod builder;
+mod data;
 mod flags;
 mod group;
 mod sampler;
@@ -96,8 +97,9 @@ use perf_event_open_sys as sys;
 use hooks::sys;
 
 pub use crate::builder::{Builder, UnsupportedOptionsError};
+pub use crate::data::{GroupData, GroupEntry, GroupIter};
 pub use crate::flags::{Clock, ReadFormat, SampleFlag, SampleSkid};
-pub use crate::group::{Group, GroupData, GroupEntry, GroupIter};
+pub use crate::group::Group;
 pub use crate::sampler::{Record, Sampler};
 
 // ... separate public exports and non-public ones
